Fix runpaths in executables.
If an executable requires a dynamic library, it'll be tagged by its
identity (id), as determined when the library was built. By default,
that'll be the path name of the library; alternatively, it can be
e.g. '@rpath/libfoo.dylib', which tells the loader to look for the
library along the 'run path' baked into the executable at build
time.

All well and good, until the suite is moved. The standard solution
does require that the whole suite, executables and dylibs, be moved as
a unit: the executable will be at $prefix/bin, so to find $prefix/lib
add the run path "@executable_path/../lib".

  * fix_executable_rpaths.sh: find the library paths relative to the
      executable, and add the rpath using @executable_path.
  * libadalang-tools.sh: fix up the executables' runpaths.
  * libadalang.sh: likewise.