Building 14.2.0, SDK16-capable.
It's not possible to build GCC (C/C++, Ada is OK) that will run on
either SDK15 or SDK16, because of header file incompatibilities that
GCC fixincludes can't cope with.

Our trick (configure with --with-specs) means that we can cope when
users have either XC or CLT installed, but it also prevents using
SDKROOT or --sysroot= at compile time.

So, this build uses a special base compiler built WITHOUT the
--with-specs trick, against SDK15(.3); and sets the built compiler to
run against $SDKROOT/../MacOSK14.sdk (for Sonoma) instead of the plain
SDKROOT (which for SDK16 builds against the problematic MacOSX15.sdk).

  * common.h:
    (BUILD): set to $ARCH-apple-darwin23
    (MACOSX_DEPLOYMENT_TARGET): set to 14 (Sonoma).
    (GCC_SRC=$SRC_PATH/gcc-14-branch): commentary notes it's
      gcc-14-2-darwin-r2.
  * gcc.sh: don't install the shim. Don't fix up rpaths.
    (XCODE): use MacOSX14.sdk.
    (CLT): likewise.
    (configure):
      (--with-build-sysroot): remove.
      (--with-sysroot): new. Use $SDKROOT/../MacOSX14.sdk.
      (--with-specs): change '%{!sysroot=*:' to '%{!sysroot:' (still "works",
        but doesn't help with runtime SDKROOT/--sysroot= problem).
      (CC): new, picks up new base compiler.
      (CXX): likewise.