Building gcc-15-20241110-x86_64.
This change was needed because the only GCC 14 x86_64 compiler
available was built with SDK14, and couldn't build the new compiler
against SDK 15 (which SDKROOT points to, for the new compiler).

The base compiler was build against SDK14 using the 'clt' branch.

CC and CXX are the C & C++ compilers respectively, used to build stage
1.

  * common.sh (SNAPSHOT): set to gcc-15-20241100.
  * gcc.sh:
    (BASE_COMPILER): new.
    (BASE_SYSROOT): new.
    (CC): formed from BASE_COMPILER and BASE_SYSROOT.
    (CXX): likewise.